modules/Single/TexRank/requirements.txt
aspose-words
breadability==0.1.20
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.0
ci-info==0.3.0
click==8.1.3
configobj==5.0.6
configparser==5.2.0
docopt==0.6.2
docx==0.2.4
etelemetry==0.3.0
filelock==3.7.1
fitz==0.0.1.dev2
Flask==2.1.0
Flask-API==3.0.post1
Flask-Cors==3.0.10
future==0.18.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kafka-python==2.0.2
looseversion==1.0.1
lxml==4.9.1
MarkupSafe==2.1.1
networkx==2.8.5
nibabel==4.0.1
nipype==1.8.3
nltk==3.7
numpy==1.23.1
packaging==21.3
pandas==1.4.3
pathlib==1.0.1
Pillow==9.2.0
prov==2.0.0
psutil==5.9.2
pycountry==22.3.5
pydot==1.4.2
pymongo==4.2.0
PyMuPDF==1.20.1
pynvml==11.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
pyxnat==1.4
rdflib==6.2.0
regex==2022.8.17
requests==2.28.1
scipy==1.9.0
simplejson==3.17.6
six==1.16.0
sumy==0.10.0
tqdm==4.64.1
traits==6.3.2
urllib3==1.26.11
Werkzeug==2.2.1
zipp==3.8.1
python-dotenv==1.0.0
